move mol_ids functions to mob_kmc
